moved App component and remove default create-react-app stuff
