+++ src/components/App.js
@@ -1,8 +1,7 @@
 import React, { Component } from 'react';
 import snoowrap from 'snoowrap';
-import logo from './logo.svg';
-import './App.css';
-import fetchAnonymousToken from './helpers';
+import '../App.css';
+import fetchAnonymousToken from '../helpers';
 
 class App extends Component {
   state = {
@@ -56,18 +55,7 @@ class App extends Component {
   }
 
   render() {
-    return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <h1 className="App-title">Welcome to React</h1>
-        </header>
-        <p className="App-intro">
-          To get started, edit <code>src/App.js</code> and save to reload.
-        </p>
-        {this.state.links}
-      </div>
-    );
+    return <div className="App">{this.state.links}</div>;
   }
 }
